Backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-jose[cryptography]
passlib[bcrypt]
pydantic
email-validator
pytest
httpx
python-multipart
python-dotenv